Changed git structure.
For now the structure will be the following:

master branch will hold the patches directory and the README.

upstream branch will keep a relatively up-to-date clone of the upstream
branch.

mine will be the branch I use to compile my st. It will be composed of
the upstream branch and all the patches applied and configurations
defined.